requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
python-dotenv
genkit
genkit-plugin-google-genai
genkit-plugin-flask @ git+https://github.com/firebase/genkit#subdirectory=py/plugins/flask